requirements.txt
duonlabs
    requests
    python-dotenv
    Pillow
    